[Server] 예외처리 추가 및 쿼리메서드 리팩토링

+++ server/src/main/java/kr/codesquad/library/domain/book/response/BooksByCategoryResponse.java
@@ -10,11 +10,11 @@ public class BooksByCategoryResponse {
 
     private final Long categoryId;
     private final String categoryTitle;
-    private final Long bookCount;
+    private final int bookCount;
     private final List<BookResponse> books;
 
     @Builder
-    public BooksByCategoryResponse(Long categoryId, String categoryTitle, Long bookCount, List<BookResponse> books) {
+    public BooksByCategoryResponse(Long categoryId, String categoryTitle, int bookCount, List<BookResponse> books) {
         this.categoryId = categoryId;
         this.categoryTitle = categoryTitle;
         this.bookCount = bookCount;
